refactor: moved Client methods related to keypackage in a dedicated mod
